Update ADC triggers (#94)
The code has been tested.

Closed and merging.

+++ zephyr/modules/owntech_spin_api/zephyr/src/data/data_conversion.cpp
@@ -45,7 +45,7 @@ static conversion_type_t conversion_types[ADC_COUNT][CHANNELS_PER_ADC];
 static float32_t* conversion_parameters[ADC_COUNT][CHANNELS_PER_ADC];
 
 #define VREF 2.048f // voltage reference from ADC
-#define QUANTUM_MAX 4096.0f // ADC resolution 
+#define QUANTUM_MAX 4096.0f // ADC resolution
 #define Vin_dividor 3.3f // Input voltage in the voltage divider
 
 /////
@@ -124,14 +124,14 @@ float32_t data_conversion_convert_raw_value(uint8_t adc_num, uint8_t channel_num
 		case conversion_therm:
 		{
 			/* Retrieves the parameters for the thermo resistor */
-			float32_t local_r0 = conversion_parameters[adc_index][channel_index][0]; 
+			float32_t local_r0 = conversion_parameters[adc_index][channel_index][0];
 			float32_t local_b = conversion_parameters[adc_index][channel_index][1];
 			float32_t local_rdiv = conversion_parameters[adc_index][channel_index][2];
 			float32_t local_t0 = conversion_parameters[adc_index][channel_index][3];
 
 			float32_t V_adc = (raw_value/QUANTUM_MAX)*VREF;							/* converts raw values into voltage */
 			float32_t R_t = (V_adc/(Vin_dividor - V_adc))*local_rdiv;				/* uses a bridge divider equation to estimate the sensor resistance */
-			float32_t T = local_t0/( 1 + log(R_t/local_r0) * (local_t0/local_b));	/* original equation R = exp(B*(1/T - 1/T0)) */
+			float32_t T = local_t0/( 1 + (float32_t)log(R_t/local_r0) * (local_t0/local_b));	/* original equation R = exp(B*(1/T - 1/T0)) */
 			return (T - 273.15f); 													/* returns value in degree Celsius */
 			break;
 		}			break;
